requirements.txt
altair==4.1.0
astor==0.8.1
backports.zoneinfo==0.2.1
base58==2.1.0
blinker==1.4
cachetools==4.2.4
click==7.1.2
ipywidgets==7.6.5
jupyterlab-widgets==1.0.2
numpy==1.21.2
pandas==1.3.3
patsy==0.5.2
Pillow==8.3.2
plotly==5.3.1
protobuf==3.18.1
pyarrow==5.0.0
pydeck==0.7.0
pytz==2021.3
scipy==1.7.1
statsmodels==0.13.0
streamlit==1.0.0
tenacity==8.0.1
toolz==0.11.1
tzlocal==3.0
validators==0.18.2
watchdog==2.1.6
widgetsnbextension==3.5.1
